renamed index->Index, added App-test
