+++ App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AuthProvider } from './src/context/AuthContext';
-import Index from './src';
+import Index from './src/Index';
 
 const App = () => {
   return (
